+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added support for query `NodeInfo`
 - Added support for query `JoinNetwork`
 - Added support for query `LeaveNetwork`
+- Added support for query `GetInstances`
 - Added support for query `GetAccountList`
 - Added support for query `GetBannedPeers`
 - Added support for query `GetBranches`